requirements.txt
# Core ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API and Web
requests>=2.31.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database
python-dotenv>=1.0.0

# Development
jupyter>=1.0.0
